// src/app/models/customer.interface.ts
export interface CustomerI {
    idCustomer?: string;
    name?: string;
    lastName?: string;
    email?: string;
    idcity?: string;
    idgender?: string;
    customerUser?: string;
    password?: string;
    cellphone?: string;
    birthday?: Date;
    accept?: boolean;
    login?: any; 
  }
  